modified:   README.md 	modified:   README_CN.md 	modified:   default_config.cfg 	modified:   src/core/basic_reduction.py 	deleted:    src/core/grating_trace.py 	deleted:    src/core/grating_trace_simple.py 	modified:   src/core/order_tracing.py 	modified:   src/core/processing_pipeline.py 	modified:   src/gui/main_window.py 	modified:   src/gui/settings_dialog.py 	modified:   src/utils/fits_io.py 	modified:   src/utils/overscan.py

+++ src/utils/overscan.py
@@ -512,7 +512,11 @@ class OverscanCorrector:
         if image.shape != bias.shape:
             raise ValueError(f"Image shape {image.shape} != bias shape {bias.shape}")
 
-        corrected = image.astype(float) - bias
+        # Explicitly convert to a high-precision float to avoid integer overflow/wraparound issues.
+        # Raw FITS data is often uint16. Directly subtracting a float bias from it can lead to
+        # incorrect results if not handled carefully. Converting to float64 is the safest path.
+        corrected = image.astype(np.float64) - bias.astype(np.float64)
+
         logger.debug(f"Overscan correction applied: "
                     f"image range [{np.min(corrected):.1f}, {np.max(corrected):.1f}]")
